+++ crates/ln-dlc-node/src/fee_rate_estimator.rs
@@ -1,21 +1,15 @@
 use anyhow::Result;
 use bdk::FeeRate;
+use bitcoin::Network;
 use lightning::chain::chaininterface::ConfirmationTarget;
 use lightning::chain::chaininterface::FeeEstimator;
 use lightning::chain::chaininterface::FEERATE_FLOOR_SATS_PER_KW;
 use parking_lot::RwLock;
 use std::collections::HashMap;
 
-const CONFIRMATION_TARGETS: [(ConfirmationTarget, usize); 4] = [
-    // We choose an extremely high background confirmation target to avoid force-closing channels
-    // unnecessarily.
-    (ConfirmationTarget::Background, 1008),
-    // We just want to end up in the mempool eventually.  We just set the target to 1008
-    // as that is esplora's highest block target available
-    (ConfirmationTarget::MempoolMinimum, 1008),
-    (ConfirmationTarget::Normal, 6),
-    (ConfirmationTarget::HighPriority, 3),
-];
+// We just want to end up in the mempool eventually.  We just set the target to 1008
+// as that is esplora's highest block target available
+const HIGHES_ALLOWED_FEE_RATE: usize = 1008;
 
 /// Default values used when constructing the [`FeeRateEstimator`] if the fee rate sever cannot give
 /// us up-to-date values.
@@ -28,7 +22,7 @@ const FEE_RATE_DEFAULTS: [(ConfirmationTarget, u32); 3] = [
 ];
 
 pub struct FeeRateEstimator {
-    client: esplora_client::BlockingClient,
+    client: mempool::MempoolFeeRateEstimator,
     fee_rate_cache: RwLock<HashMap<ConfirmationTarget, FeeRate>>,
 }
 
@@ -42,20 +36,40 @@ impl EstimateFeeRate for FeeRateEstimator {
     }
 }
 
+fn to_mempool_network(value: Network) -> mempool::Network {
+    match value {
+        Network::Bitcoin => mempool::Network::Mainnet,
+        Network::Testnet => mempool::Network::Testnet,
+        Network::Signet => mempool::Network::Signet,
+        Network::Regtest => mempool::Network::Local,
+    }
+}
+
 impl FeeRateEstimator {
     /// Constructor for the [`FeeRateEstimator`].
-    pub fn new(esplora_url: String) -> Self {
-        let client = esplora_client::BlockingClient::from_agent(esplora_url, ureq::agent());
+    pub fn new(network: Network) -> Self {
+        let client = mempool::MempoolFeeRateEstimator::new(to_mempool_network(network));
 
-        let initial_fee_rates = match client.get_fee_estimates() {
-            Ok(estimates) => {
-                HashMap::from_iter(CONFIRMATION_TARGETS.into_iter().map(|(target, n_blocks)| {
-                    let fee_rate = esplora_client::convert_fee_rate(n_blocks, estimates.clone())
-                        .expect("fee rates for our confirmation targets");
-                    let fee_rate = FeeRate::from_sat_per_vb(fee_rate);
-
-                    (target, fee_rate)
-                }))
+        let initial_fee_rates = match client.fetch_fee_sync() {
+            Ok(fee_rate) => {
+                let mut hash_map = HashMap::new();
+                hash_map.insert(
+                    ConfirmationTarget::MempoolMinimum,
+                    FeeRate::from_sat_per_vb(fee_rate.minimum_fee as f32),
+                );
+                hash_map.insert(
+                    ConfirmationTarget::Background,
+                    FeeRate::from_sat_per_vb(HIGHES_ALLOWED_FEE_RATE as f32),
+                );
+                hash_map.insert(
+                    ConfirmationTarget::Normal,
+                    FeeRate::from_sat_per_vb(fee_rate.economy_fee as f32),
+                );
+                hash_map.insert(
+                    ConfirmationTarget::HighPriority,
+                    FeeRate::from_sat_per_vb(fee_rate.fastest_fee as f32),
+                );
+                hash_map
             }
             Err(e) => {
                 tracing::warn!(defaults = ?FEE_RATE_DEFAULTS, "Initializing fee rate cache with default values: {e:#}");
@@ -85,21 +99,22 @@ impl FeeRateEstimator {
     }
 
     pub(crate) async fn update(&self) -> Result<()> {
-        let estimates = self.client.get_fee_estimates()?;
+        let estimates = self.client.fetch_fee_sync()?;
 
         let mut locked_fee_rate_cache = self.fee_rate_cache.write();
-        for (target, n_blocks) in CONFIRMATION_TARGETS {
-            let fee_rate = esplora_client::convert_fee_rate(n_blocks, estimates.clone())?;
 
-            let fee_rate = FeeRate::from_sat_per_vb(fee_rate);
-
-            locked_fee_rate_cache.insert(target, fee_rate);
-            tracing::trace!(
-                n_blocks_confirmation = %n_blocks,
-                sats_per_kwu = %fee_rate.fee_wu(1000),
-                "Updated fee rate estimate",
-            );
-        }
+        locked_fee_rate_cache.insert(
+            ConfirmationTarget::MempoolMinimum,
+            FeeRate::from_sat_per_vb(estimates.minimum_fee as f32),
+        );
+        locked_fee_rate_cache.insert(
+            ConfirmationTarget::Normal,
+            FeeRate::from_sat_per_vb(estimates.economy_fee as f32),
+        );
+        locked_fee_rate_cache.insert(
+            ConfirmationTarget::HighPriority,
+            FeeRate::from_sat_per_vb(estimates.fastest_fee as f32),
+        );
 
         Ok(())
     }